Create wordsearch.py
Input a string for every row in the grid and a string for every needed word and it'll identify the row, column, positive diagonal or negative diagonal (gives the diagonal's starting position) that you'll find the word in, specifying if the word is backwards. Comments explaining the program to come.